+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Transition.hs
@@ -6,17 +6,14 @@
 
 module Cardano.Ledger.Alonzo.Transition (
   TransitionConfig (..),
-  toAlonzoTransitionConfigPairs,
 ) where
 
 import Cardano.Ledger.Alonzo.Era
 import Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis)
 import Cardano.Ledger.Alonzo.Translation ()
-import Cardano.Ledger.BaseTypes (toKeyValuePairs)
 import Cardano.Ledger.Mary
 import Cardano.Ledger.Mary.Transition (TransitionConfig (MaryTransitionConfig))
 import Cardano.Ledger.Shelley.Transition
-import Data.Aeson (KeyValue (..))
 import GHC.Generics
 import Lens.Micro
 import NoThunks.Class (NoThunks (..))
@@ -39,7 +36,3 @@ instance EraTransition AlonzoEra where
     lens atcAlonzoGenesis (\atc ag -> atc {atcAlonzoGenesis = ag})
 
 instance NoThunks (TransitionConfig AlonzoEra)
-
-toAlonzoTransitionConfigPairs :: KeyValue e a => TransitionConfig AlonzoEra -> [a]
-toAlonzoTransitionConfigPairs = toKeyValuePairs
-{-# DEPRECATED toAlonzoTransitionConfigPairs "In favor of `toKeyValuePairs`" #-}