+++ libs/cardano-ledger-core/src/Cardano/Ledger/EpochBoundary.hs
@@ -1,6 +1,0 @@
-module Cardano.Ledger.EpochBoundary
-  {-# DEPRECATED "Use `Cardano.Ledger.State` instead" #-} (
-  module Cardano.Ledger.State.SnapShots,
-) where
-
-import Cardano.Ledger.State.SnapShots