+++ eras/shelley/impl/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## 1.18.0.0
 
+* Remove deprecated function `getPoolParameters`
+* Remove deprecated function `toShelleyGenesisPairs`
+* Remove deprecated type `RewardAccounts`
+* Remove deprecated lenses `esAccountStateL`, `asTreasuryL`, `asReservesL`, `utxosUtxoL`, `epochStateTreasuryL`
+* Remove deprecated type `PPUpdateEnv`
+* Remove deprecated function `toShelleyTransitionConfigPairs`
+* Remove deprecated function `hashShelleyTxAuxData`
+* Remove deprecated function `addrWits'`
+* Remove deprecated type `Shelley`
+* Remove deprecated module `Cardano.Ledger.Shelley.BlockChain`
 * Change state of `ShelleyDELEG` rule from `DState` to `CertState`
 * Add `EraCertState` constraint to `STS` instance for `ShelleyDELEG`
 * Add `TxLevel` argument to `Tx` and `TxBody`
@@ -20,6 +30,10 @@
 
 ### `testlib`
 
+* Remove deprecated module `Test.Cardano.Ledger.Shelley.Generator.Constants`
+* Remove deprecated module `Test.Cardano.Ledger.Shelley.Generator.Delegation`
+* Remove deprecated module `Test.Cardano.Ledger.Shelley.Generator.Trace.DCert`
+* Remove deprecated functions `getReward`, `lookupReward`, `getRewardAccountAmount`
 * Rename `examplePoolParams` to `exampleStakePoolParams`
 * Removed `shelleyEraSpecificSpec`
 * Added `shelleyGenUnRegTxCert`